submit java-collections-ru/streams

+++ java-collections-ru/streams/src/main/java/exercise/App.java
@@ -6,8 +6,8 @@ import java.util.Arrays;
 // BEGIN
 class App {
 
-    public static Integer getCountOfFreeMails(List<String> emailsList) {
-        Integer freeDomainsNumber = emailsList.stream()
+    public static long getCountOfFreeMails(List<String> emailsList) {
+        long freeDomainsNumber = emailsList.stream()
                 .filter(str -> str.contains("@gmail.com") || str.contains("@yandex.ru") || str.contains("hotmail.com"))
                     .count();
             return freeDomainsNumber;
